% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BchAddresses.R
\name{PubHash2BtcAdr}
\alias{PubHash2BtcAdr}
\title{Create BCH address from public key hash (BTC alias)}
\usage{
PubHash2BtcAdr(pubhash)
}
\arguments{
\item{pubhash}{\code{character}, the public key hash.}
}
\value{
\code{character}, the BTC address
}
\description{
This function returns the corresponding BTC address from a
hashed public key.
}
\references{
\url{https://en.bitcoin.it/wiki/Address}
}
\author{
Bernhard Pfaff
}
\concept{BtcAdresses}
