% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitcoind.R
\name{startbtc}
\alias{startbtc}
\title{Start bitcoind server process (BTC alias)}
\usage{
startbtc(confbtc)
}
\arguments{
\item{confbtc}{\code{CONRPC} object, returned from \code{conrpc()}.}
}
\value{
\code{NULL}
}
\description{
This function is an alias of startbch
}
\details{
The process is started by calling \code{system()}.
Hereby, the options: \code{rpcuser}, \code{rpcpassword} and
\code{conf} are used in the call to \code{bitcoind}.
}
\seealso{
Other bitcoind functions: 
\code{\link{ANSRPC-class}},
\code{\link{CONRPC-class}},
\code{\link{NullOrCharacter-class}},
\code{\link{NullOrInteger-class}},
\code{\link{conrpc}()},
\code{\link{rpcpost}()},
\code{\link{startbch}()},
\code{\link{stopbch}()},
\code{\link{stopbtc}()}
}
\author{
Bernhard Pfaff
}
\concept{bitcoind functions}
