% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{validBchAdr}
\alias{validBchAdr}
\title{Validate S4-class BCHADR}
\usage{
validBchAdr(object)
}
\arguments{
\item{object}{\code{BCHADR} object}
}
\description{
This function validates objects of S4-class
\code{BCHADR}. Hereby, checks are conducted
with respect to the first character of the addresses;
their consistency with the net version and
the correspondence of the checksums.
}
\references{
\url{https://en.bitcoin.it/wiki/Address}
}
\seealso{
Other BchAdresses: 
\code{\link{BCHADR-class}},
\code{\link{BTCADR-class}},
\code{\link{PubHash2BchAdr}()},
\code{\link{PubKey2PubHash}()},
\code{\link{base58CheckDecode}()},
\code{\link{base58CheckEncode}()},
\code{\link{concatHex}()},
\code{\link{decodeHex}()},
\code{\link{hash160}()},
\code{\link{hash256}()},
\code{\link{validBtcAdr}()}
}
\author{
Bernhard Pfaff
}
\concept{BchAdresses}
